bb0: fix abv column type float→decimal(5,2) for precision

+++ database/migrations/2026_04_22_000001_create_recipes_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->text('description')->nullable();
             $table->text('instructions')->nullable();
             $table->string('glass')->nullable();
-            $table->float('abv')->nullable();
+            $table->decimal('abv', 5, 2)->nullable();
             $table->integer('volume')->nullable();
             $table->string('icon')->nullable();
             $table->string('photo')->nullable();
